Path adjustment algorithm
pathAdjust can take our computed target vector's angle and our current
path's angle, and compute a path adjustment to get us in line with where
we need to be going.
The algorithm takes the target angle and computes its alpha and beta
angles, which are +90 and +270 from the target angle respectively. Based
on whether the current angle is closer to alpha or beta, we will turn
clockwise (right) or counterclockwise (left). The amount needed to turn
is also computed.